+++ src/test/java/com/github/dockerjava/netty/exec/WaitContainerCmdExecTest.java
@@ -1,13 +1,13 @@
 package com.github.dockerjava.netty.exec;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.isEmptyString;
-import static org.hamcrest.Matchers.not;
-
-import java.lang.reflect.Method;
-
+import com.github.dockerjava.api.command.CreateContainerResponse;
+import com.github.dockerjava.api.command.InspectContainerResponse;
+import com.github.dockerjava.api.exception.DockerClientException;
+import com.github.dockerjava.api.exception.DockerException;
+import com.github.dockerjava.api.exception.NotFoundException;
+import com.github.dockerjava.api.model.WaitResponse;
+import com.github.dockerjava.core.command.WaitContainerResultCallback;
+import com.github.dockerjava.netty.AbstractNettyDockerClientTest;
 import org.testng.ITestResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.AfterTest;
@@ -15,13 +15,11 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-import com.github.dockerjava.api.command.CreateContainerResponse;
-import com.github.dockerjava.api.command.InspectContainerResponse;
-import com.github.dockerjava.api.exception.DockerException;
-import com.github.dockerjava.api.exception.NotFoundException;
-import com.github.dockerjava.api.model.WaitResponse;
-import com.github.dockerjava.core.command.WaitContainerResultCallback;
-import com.github.dockerjava.netty.AbstractNettyDockerClientTest;
+import java.lang.reflect.Method;
+import java.util.concurrent.TimeUnit;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
 
 @Test(groups = "integration")
 public class WaitContainerCmdExecTest extends AbstractNettyDockerClientTest {
@@ -75,7 +73,7 @@ public class WaitContainerCmdExecTest extends AbstractNettyDockerClientTest {
         WaitContainerResultCallback callback = new WaitContainerResultCallback() {
             public void onNext(WaitResponse waitResponse) {
                 fail("expected NotFoundException");
-            };
+            }
         };
 
         dockerClient.waitContainerCmd("non-existing").exec(callback).awaitStatusCode();
@@ -106,5 +104,25 @@ public class WaitContainerCmdExecTest extends AbstractNettyDockerClientTest {
         LOG.info("Container Inspect: {}", inspectContainerResponse.toString());
 
         assertThat(inspectContainerResponse.getState().getRunning(), is(equalTo(false)));
+    }
+
+    @Test
+    public void testWaitContainerTimeout() throws Exception {
+
+        CreateContainerResponse container = dockerClient.createContainerCmd("busybox").withCmd("sleep", "10").exec();
+
+        LOG.info("Created container: {}", container.toString());
+        assertThat(container.getId(), not(isEmptyString()));
+
+        dockerClient.startContainerCmd(container.getId()).exec();
+
+        WaitContainerResultCallback callback = dockerClient.waitContainerCmd(container.getId()).exec(
+                new WaitContainerResultCallback());
+        try {
+            callback.awaitStatusCode(100, TimeUnit.MILLISECONDS);
+            fail("Should throw exception on timeout.");
+        } catch(DockerClientException e){
+            LOG.info(e.getMessage());
+        }
     }
 }